+++ src/reducers.js
@@ -13,6 +13,8 @@ import selectedMoveId from 'scenes/Game/reducers/selectedMoveId';
 // Board
 import selectedSlot from 'components/Board/reducers/selectedSlot';
 
+import { Color } from 'services/logic/constants';
+
 const rootReducer = combineReducers({
   // Home,
   autoMove,
@@ -50,3 +52,46 @@ export const getNextMove = ({ moves, selectedMoveId }) =>
 
 export const getNextMoveColor = ({ moves }) =>
   fromMoves.getNextMoveColor(moves);
+
+/*****************/
+/***  Players  ***/
+/*****************/
+
+const lookupPlayer = (players, key, value) =>
+  players.find(p => p[key] === value);
+
+export const getNextMovePlayer = ({ players, moves }) =>
+  lookupPlayer(players, 'color', getNextMoveColor({ moves }));
+
+const getUserPlayer = ({ players, username }) =>
+  lookupPlayer(players, 'name', username);
+
+export const getRedPlayer = ({ players }) =>
+  lookupPlayer(players, 'color', Color.RED);
+
+export const getBlackPlayer = ({ players }) =>
+  lookupPlayer(players, 'color', Color.BLACK);
+
+export const getUserColor = ({ players, username }) => {
+  try {
+    return getUserPlayer({ players, username }).color;
+  } catch (e) {
+    if (e instanceof TypeError) return undefined;
+    throw e;
+  }
+};
+
+export const getOtherPlayer = ({ gameSlug, players, username }) => {
+  if (gameSlug === null) getBlackPlayer({ players });
+  return players.find(p => p.name !== username);
+};
+
+// TODO: add a state that allows players to flip their original orientation
+export const getInitialUserOrientation = ({ players, username }) =>
+  getUserColor({ players, username }) === Color.BLACK;
+
+// TODO: move to layout class that displays board and players
+export const getCurrentPlayer = ({ gameSlug, players, username }) => {
+  if (gameSlug === null) getRedPlayer({ players });
+  return getUserPlayer({ players, username });
+};
