+++ src/scenes/Game/selectors.js
@@ -1,53 +1,9 @@
-import { Color } from 'services/logic/constants';
-
 import {
   getLastMove,
   getSelectedMove,
   getNextMoveColor,
+  getUserColor,
 } from 'reducers';
-
-/*****************/
-/***  Players  ***/
-/*****************/
-
-const lookupPlayer = (players, key, value) =>
-  players.find((p) => p[key] === value);
-
-export const getNextMovePlayer = ({ players, moves }) =>
-  lookupPlayer(players, 'color', getNextMoveColor({ moves }));
-
-const getUserPlayer = ({ players, username }) =>
-  lookupPlayer(players, 'name', username);
-
-export const getRedPlayer = ({ players }) =>
-  lookupPlayer(players, 'color', Color.RED);
-
-export const getBlackPlayer = ({ players }) =>
-  lookupPlayer(players, 'color', Color.BLACK);
-
-export const getUserColor = ({ players, username }) => {
-  try {
-    return getUserPlayer({ players, username }).color;
-  } catch (e) {
-    if (e instanceof TypeError) return undefined;
-    throw e;
-  }
-};
-
-export const getOtherPlayer = ({ gameSlug, players, username }) => {
-  if (gameSlug === null) getBlackPlayer({ players });
-  return players.find((p) => p.name !== username);
-};
-
-// TODO: add a state that allows players to flip their original orientation
-export const getInitialUserOrientation = ({ players, username }) =>
-  getUserColor({ players, username }) === Color.BLACK;
-
-// TODO: move to layout class that displays board and players
-export const getCurrentPlayer = ({ gameSlug, players, username }) => {
-  if (gameSlug === null) getRedPlayer({ players });
-  return getUserPlayer({ players, username });
-};
 
 /********************/
 /***  Game Logic  ***/
@@ -74,7 +30,7 @@ export const getLegalMoves = (
 };
 
 // TODO: this isn't really a selector -- use refactor with mapStateToProps
-export const hasLegalMoves = (state) => {
+export const hasLegalMoves = state => {
   const { board } = getLastMove(state);
   return board.hasLegalMoves(getNextMoveColor(state));
 };
